GlobalDevelopers.DataCaptureSolutions/GlobalDevelopers.DataCaptureSolutions/Controllers/HomeController.cs
using System.Web.Mvc;

namespace GlobalDevelopers.DataCaptureSolutions.Controllers
{
    public class HomeController : Controller
    {
        public ActionResult Index()
        {
            return View();
        }

        public ActionResult AlternateStyle()
        {
            ViewBag.Message = "Your application description page.";

            return View();
        }

        public ActionResult Contact()
        {
            ViewBag.Message = "Your contact page.";

            return View();
        }
    }
}